requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
backports.csv==1.0.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
bloom-filter==1.3
boto==2.49.0
boto3==1.14.58
botocore==1.17.58
bs4==0.0.1
celery==4.4.7
certifi==2019.6.16
cffi==1.14.2
channels==2.4.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
cv==1.0.0
Cython==0.29.14
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-celery==3.3.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.0.0
django-import-export==1.2.0
django-ranged-response==0.2.0
django-redis==4.11.0
django-reversion==3.0.7
django-silk==4.1.0
django-simple-captcha==0.5.12
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
docutils==0.15.2
drf-extensions==0.6.0
et-xmlfile==1.0.1
eventlet==0.26.1
future==0.17.1
fuzzywuzzy==0.18.0
gensim==3.8.3
gevent==20.6.2
gprof2dot==2019.11.30
greenlet==0.4.16
httplib2==0.13.1
hyperlink==20.0.1
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
iniconfig==1.0.1
itypes==1.1.0
jdcal==1.4.1
jieba==0.42.1
Jinja2==2.10.1
jmespath==0.10.0
kombu==4.6.11
lxml==4.5.2
Markdown==3.2.1
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==8.4.0
msgpack-python==0.5.6
mysqlclient==1.4.2.post1
numpy==1.19.1
odfpy==1.4.0
opencv-python==4.4.0.40
openpyxl==2.6.2
packaging==20.4
pandas==1.1.0
pika==1.1.0
Pillow==7.0.0
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.7
pydub==0.24.1
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.0.1
python-dateutil==2.8.1
pytz==2019.1
PyYAML==5.1.1
raven==6.10.0
redis==3.4.1
reportlab==3.5.47
requests==2.22.0
reversion==0.2
s3transfer==0.3.3
scipy==1.5.2
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
smart-open==2.1.1
soupsieve==2.0.1
sqlparse==0.3.0
tablib==0.13.0
tencent-cloud-sdk==0.2.2
tencent-cloud-sdk-auth==0.2.1
tencent-cloud-sdk-common==0.1.1
tencent-cloud-sdk-core==0.2.1
tencent-cloud-sdk-serverless-database==0.1.1
tencent-cloud-sdk-serverless-functions==0.2.1
tencentcloud-sdk-python==3.0.253
toml==0.10.0
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
websocket==0.2.1
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
